Double values

+++ Game/RandGen.h
@@ -8,8 +8,8 @@ class RandGen
 {
 private:
 //	int N, Prob, ES, ET, EG, AS, AM, AD;
-	int A,B;
-	int P, H, AC,T;
+	double A,B;
+	double P, H, AC,T;
 public:
 	RandGen(int n, int p, int Ep[], int Ap[],int ER[], int AR[],int TS,EarthArmy* E);
 	unit* GenEarth(int Ep[], int ER[]);
